AI Speech/requirements.txt
azure-cognitiveservices-speech==1.43.0
load-dotenv==0.1.0
keyboard